Move helpers to a separate file; fix issue with showing time estimate when fullSyncHeight is missing
